fix : GameTest 오류 수정, refactor : TrialTimeTest 클래스 생성
