+++ src/test/java/racingcarTest/TrialTimeTest.java
@@ -2,57 +2,25 @@ package racingcarTest;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.Model.Name;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.CsvSource;
 import racingcar.Model.TrialTime;
 
 import java.util.InputMismatchException;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-/**
- * 클래스 이름 : inputValidationTest.java
- *
- * @author 보스독
- * @author 작은곰
- * @author 토니
- * @version 1.0
- * <p>
- * 날짜 : 2020.02.13 목요일
- */
-
-public class InputValidationTest {
-    @Test
-    @DisplayName("차 이름에 빈 문자열 입력 시")
-    void inputValidation_차_이름_빈_문자열() {
-        String carNames = "";
-        assertThatThrownBy(() -> {
-            Name.checkEmptyCarName(carNames);
-        }).isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("차 이름은 빈 문자열일 수 없습니다.");
-
-    }
-
-    @Test
-    @DisplayName("차 이름이 5글자 이내")
-    void inputValidation_차_이름_글자수() {
-        assertThatThrownBy(() -> {
-            String result = "Maserati";
-            Name.checkSmallerThanSix(result);
-        }).isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("차 이름의 길이가 6 이상입니다.");
-    }
-
-    @Test
+public class TrialTimeTest {
+    @ParameterizedTest
+    @CsvSource(value = {"-4", "0"})
     @DisplayName("시도횟수가 음수 또는 0")
-    void inputValidation_시도횟수가_0이하() {
+    void inputValidation_시도횟수가_0이하(String trialTime) {
         assertThatThrownBy(() -> {
-            String trialTime = "-4";
             TrialTime.checkNegativeAndZeroInput(trialTime);
         }).isInstanceOf(IllegalArgumentException.class)
                 .hasMessage("음수 또는 0을 입력할 수 없습니다.");
 
         assertThatThrownBy(() -> {
-            String trialTime = "0";
             TrialTime.checkNegativeAndZeroInput(trialTime);
         }).isInstanceOf(IllegalArgumentException.class)
                 .hasMessage("음수 또는 0을 입력할 수 없습니다.");
